+++ config/description.exs
@@ -765,6 +765,15 @@ config :pleroma, :config_description, [
             " This will break federation with older software for theses nicknames."
       },
       %{
+        key: :cleanup_attachments,
+        type: :boolean,
+        description: """
+        "Set to `true` to remove associated attachments when status is removed.
+        This will not affect duplicates and attachments without status.
+        Enabling this will increase load to database when deleting statuses on larger instances.
+        """
+      },
+      %{
         key: :max_pinned_statuses,
         type: :integer,
         description: "The maximum number of pinned statuses. 0 will disable the feature.",